==> Code/0 - Scraping and Cleaning.ipynb <==
!pip install tweepy
!pip install praw

# Import packages
import praw
import tweepy
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
import time
import pandas as pd
import datetime as dt
import csv
import psycopg2
import yfinance as yf
from functools import reduce

#Filtering funciton that cleans each set and adds price column
def filter_df(df, data_type, coin_name, coin_ticker):
    
    df = df.copy()
    
    df['date_scrape'] = pd.to_datetime(df['date_scrape'])
   

    if data_type == 'reddit':
        df['subreddit'] = df['subreddit'].str.lower()
        filtered_df = df[df['subreddit'] == coin_name]

    if data_type == 'twitter':
        df['user_name'] = df['user_name'].str.lower()
        filtered_df = df[df['user_name'] == coin_name]

    if data_type == 'youtube':
        df['search_word'] = df['search_word'].str.lower()
        filtered_df = df[df['search_word'] == coin_name]
        
    

    # Query price data
    coin_df = yf.download(coin_ticker, start=df['date_scrape'].min(), end=df['date_scrape'].max()).reset_index().drop(columns=['Open', 'High', 'Low', 'Adj Close', 'Volume'])
    coin_df = coin_df.rename(columns = {'Date': 'date_scrape'})

    # Merge 
    output_df = filtered_df.merge(coin_df, on='date_scrape', how = 'outer')
    output_df = output_df.ffill()
    
    #Making sure dates are in order        
    output_df = output_df.sort_values(by = ['date_scrape']).reset_index() 
    output_df = output_df.drop(['index'], axis=1)

    return output_df

## Reddit Scraping

# Access Reddit API 
reddit = praw.Reddit(client_id= "ibh9FWjQqVYZYw",                      # your client id
                     client_secret= "VeGeOHgpCN9SVSDsClpaR_IRWsCfiA",  # your client secret
                     user_agent= "reddit_scraper",                     # user agent (app) name
                     username = "asharm2949",                          # your reddit username asharm2949
                     password = "aman-sharma2949",                     # your reddit password
                     check_for_async=False)               

# Subreddit name and queries list to srape
scrape_dict = {
                'Bitcoin': ['btc', 'bitcoin'],
                'ethereum': ['eth', 'ethereum'],
                'litecoin': ['ltc', 'litecoin'],
                'algorand': ['algo', 'algorand'],
                'cardano': ['ada', 'cardano'],
                'Tether': ['usdt', 'tether'],
                'Ripple': ['xrp', 'ripple'],
                'cosmosnetwork': ['atom', 'cosmos'],
                'dot': ['dot', 'polkadot'],
                'dogecoin': ['doge', 'dogecoin']
}

scrape_list = ['CryptoCurrency']

# Helper function 
def get_date(created):
    return dt.datetime.fromtimestamp(created)

def scrape_reddit_mass(scrape_list):
  
  # Post data dictionary
  post_dict = {
      "timestamp_scrape" : [],  # timestamp of the scrape
      "date_scrape" : [],       # date of the scrape    
      "timestamp_post" : [],    # timestamp of the post
      "subreddit": [],          # subreddit of the post
      "id" : [],                # unique id of the post
      "title" : [],             # title of the post
      "url" : [],               # url of the post
      "body" : [],              # the descriptionof post
      "score" : [],             # number of upvotes of the post
      "upvote_ratio": [],       # ratio of upvotes to total votes
      "num_comms": [],          # the number of comments on the post
  }

  for subreddit in scrape_list:
    posts = reddit.subreddit('CryptoCurrency').hot(limit=5000)
    
    for post in posts:
      post_dict['timestamp_scrape'].append(pd.to_datetime("now"))
      post_dict['date_scrape'].append(dt.date.today())
      post_dict["timestamp_post"].append(post.created_utc)
      post_dict["subreddit"].append(post.subreddit)
      post_dict["id"].append(post.id)
      post_dict["title"].append(post.title)
      post_dict["url"].append(post.url)
      post_dict["body"].append(post.selftext)
      post_dict["score"].append(post.score)
      post_dict["upvote_ratio"].append(post.upvote_ratio)           
      post_dict["num_comms"].append(post.num_comments)

  reddit_data_mass = pd.DataFrame(post_dict)
  reddit_data_mass['timestamp_post'] = reddit_data_mass["timestamp_post"].apply(get_date)

  return reddit_data_mass

## Twitter Scraper

# Access Twitter API
CONSUMER_KEY = 'PqIOkmsFKDc8lIoSaDGf0Kcr0'
CONSUMER_SECRET = '50kaIGjkmKSSFTKoFtte8j6vbR0CX7KPUDqB4oRhpW0k7zPuY4'
ACCESS_TOKEN = '1347086178365829123-nDPxqMJfjuf9qgVqOUQ8A8LsfZw7pD'
ACCESS_TOKEN_SECRET = 'gT0tbfA9Ca97GbyAwODU8W453sSxLPpNwv800PFpvnXYo'
# Bearer Token = 'AAAAAAAAAAAAAAAAAAAAAPA%2BMwEAAAAAYYSYRh%2B8n9rbJaxva3Y%2FWqtOidU%3D5wukphfJfVhYaxG6HYIPF7nbEI0f5ShwXgs3U8ccIhTC16rSxL'

auth = tweepy.OAuthHandler(CONSUMER_KEY, CONSUMER_SECRET)
auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)

api = tweepy.API(auth, wait_on_rate_limit=True)

# Function to scrape twitter
def scrape_twitter(search_word, num_tweets):
    
  # Scrape tweets
  tweets = tweepy.Cursor(api.search,
                          q=search_word + " -filter:retweets",
                          wait_on_rate_limit = True,
                          wait_on_rate_limit_notify = True,
                          lang="en",
                          count=num_tweets,
                          since=dt.date.today()).items()

  # Post data dictionary
  post_dict = {
      
    "timestamp_scrape" : [],  # timestamp of the scrape
    "date_scrape" : [],       # date of the scrape
    "timestamp_tweet" : [],   # timestamp of the tweet
    "search_word": [],        # requested query 
    "tweet_id" : [],          # id of tweet
    "user_name": [],          # name of user who tweeted
    "user_location": [],      # location of user 
    "text" : [],              # text of the tweet
    "retweet_count": [],      # number of retweets
    "favorite_count": []      # number of likes
  }

  # Collect a list of tweets
  for tweet in tweets:
    post_dict['timestamp_scrape'].append(pd.to_datetime("now"))
    post_dict['date_scrape'].append(dt.date.today())
    post_dict["timestamp_tweet"].append(tweet.created_at)
    post_dict["search_word"].append(search_word)
    post_dict["tweet_id"].append(tweet.id)    
    post_dict["user_name"].append(tweet.user.screen_name)       
    post_dict["user_location"].append(tweet.user.location)   
    post_dict["text"].append(tweet.text)
    post_dict["retweet_count"].append(tweet.retweet_count)
    post_dict["favorite_count"].append(tweet.favorite_count)

  twitter_data = pd.DataFrame(post_dict)
  return twitter_data

## Youtube Scraper

# Youtube queries to search
youtube_query_list = ['stellar crypto', 'ripple', 'cosmos crypto', 'dot crypto', 'dogecoin',
                      'bitcoin', 'ethereum', 'litecoin', 'algorand', 'cardano',
                      'tezos', 'monero', 'dai crypto', 'filecoin', 'tron crypto',
                      'eos crypto', 'chainlink']

# Helper function (format numeric fields)
def clean_youtube_data(input_df):

    watching = pd.DataFrame(input_df['views'].str.find('watching'))
    mask = (watching['views'] == -1)
    df = input_df[mask].reset_index(drop = True)

    for index, row in df.iterrows():
        try:
            # channel_subs
            subs = row['channel_subs']
            index_w = subs.find(" ")
            subs = subs[0:index_w]
            real_subs = 0

            if (subs[len(subs)-1] == "M"):
                real_subs = int(1000000*float(subs[0:len(subs)-1]))

            elif (subs[len(subs)-1] == "K"):
                real_subs = int(1000*float(subs[0:len(subs)-1]))

            else:
                real_subs = int(subs)

            df.iat[index, 4] = real_subs

        except:
            pass
        
        try:
            # channel name
            channel_name = row['channel_name']
            channel_name = channel_name.replace('\n', '')
            channel_name = channel_name.strip()
            df.iat[index, 3] = channel_name

        except:
            pass
        
        try:
            # title
            title = row['title']
            title = title.replace('\n', '')
            title = title.strip()
            df.iat[index, 5] = title

        except:
            pass
        
        try:
            # date posted
            date_posted = row['date_posted']
            date_posted = date_posted.split()
            date_posted = date_posted[-3] + " " + date_posted[-2] + " " + date_posted[-1]
            date_posted = dt.datetime.strptime(date_posted, '%b %d, %Y').date()
            df.iat[index, 6] = date_posted

        except:
            pass

        try:
            # num_views
            views = row['views']

            if ('watching' in views):
                df = df.drop(index)
                continue

            index_w2 = views.find(" ")
            real_views = views[0:index_w2]
            real_views = int(real_views.replace(",", ''))
            df.iat[index, 9] = real_views

        except:
            pass

        try:
            # num_likes
            likes = row['num_likes']
            likes = likes.rstrip(' ')
            index_w3 = likes.find("M")
            index_w4 = likes.find("K")

            if (index_w3 != -1):
                likes = float(likes[0:index_w3])*1000000

            elif (index_w4 != -1):
                likes =float(likes[0:index_w4])*1000

            df.iat[index, 11] = likes

        except:
            pass

        # num_dislikes
        try:
            dislikes = row['num_dislikes']
            index_w5 = dislikes.find("M")
            index_w6 = dislikes.find("K")

            if (index_w5 != -1):
                dislikes = float(dislikes[0:index_w5])*1000000

            elif (index_w6 != -1):
                dislikes = float(dislikes[0:index_w6])*1000
            df.iat[index, 12] = dislikes
        except:
            pass

        # num_comments
        try:
            comments = row['num_comments']
            comments = int(comments.replace(",",""))
            df.iat[index, 10] = comments

        except:
            pass
        
        try:
            # description
            description = row['description']
            description = description.replace('\n', '')
            description = description.strip()
            df.iat[index, 8] = description

        except:
            pass


    return df

# Scrape function (coin queries)
def scrape_youtube(search_query_list, num_vids):
    # Access URL to scrape
    template_url = 'https://www.youtube.com/results?search_query={}&sp=CAMSBAgEEAE%253D'

    # Post data dictionary
    post_dict = {
            "timestamp_scrape" : [],  # timestamp of scrape
            "date_scrape": [],        # date of scrape
            "search_word" : [],       # search query
            "channel_name" : [],      # name of posting account
            "channel_subs" : [],      # number of subscribers on posting account
            "title": [],              # title of video
            "date_posted" : [],       # date video was posted
            "url": [],                # title of video
            "description": [],        # description of video
            "views": [],              # number of video views
            "num_comments": [],       # number of video views
            "num_likes": [],          # number of video likes
            "num_dislikes": []        # number of video dislikes
    }

    chrome_options = Options()
    chrome_options = webdriver.ChromeOptions()
    chrome_options.add_argument("--disable-notifications")
    #chrome_options.add_argument("--headless")
    #chrome_options.add_argument("--disable-gpu")
    #chrome_options.add_argument('--no-sandbox')
    #chrome_options.add_argument('--disable-dev-shm-usage')
    #chrome_options.add_argument("--window-size=1920,1920")
    #chrome_options.add_argument("--mute-audio")


    for search_query in search_query_list:
        url = template_url.format(search_query)
        PATH = "/Users/ArpanBagui/Documents/UPenn/Projects/CryptoDash/chromedriver"
        #driver = webdriver.Chrome(executable_path=PATH, options=chrome_options)
        driver = webdriver.Chrome(ChromeDriverManager().install())
        #driver = webdriver.Chrome(options=chrome_options)
        driver.get(url)
        time.sleep(2)

        # Find ads in results
        path = '//*[@id="ad-badge-container"]/ytd-badge-supported-renderer/div/span'
        ad_list= driver.find_elements_by_xpath(path)
        num_ads = len(ad_list)
        
        #Variable that counts vid num
        vid_count = 0
        
        for video in driver.find_elements_by_id('video-title')[num_ads:num_vids + num_ads]:
            #Adding to vid count
            vid_count += 1
            
            try:
                video.click()
                time.sleep(5)
            except:
                continue
            
            page_source = driver.page_source
            soup = BeautifulSoup(page_source, 'lxml')
                
            try:
                post_dict['url'].append(video.get_attribute('href'))
            except StaleElementReferenceException:
                post_dict['url'].append('')

            try:
                post_dict['title'].append(video.get_attribute('text'))
            except StaleElementReferenceException:
                post_dict['title'].append('')

            try:
                #Variable that allows us to skip over previous channels (weird error with soup)
                channel_index = vid_count*2 - 1
                channel = soup.find_all('a', {"class" : "yt-simple-endpoint style-scope yt-formatted-string"})[channel_index].get_text()
                post_dict['channel_name'].append(channel)

            except TimeoutException:
                post_dict['channel_name'].append('')

            try:
                subs = soup.findAll('yt-formatted-string', id = "owner-sub-count")[0].get_text()
                post_dict['channel_subs'].append(subs)

            except TimeoutException:
                post_dict['channel_subs'].append('')

                
            try:
                date_posted = soup.find_all('yt-formatted-string', {"class" : "style-scope ytd-video-primary-info-renderer"})[1].get_text()
                post_dict['date_posted'].append(date_posted)

            except TimeoutException:
                post_dict['date_posted'].append('')

            
            try:
                WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.CLASS_NAME, 'view-count')))
                post_dict['views'].append(driver.find_element_by_class_name('view-count').text)

            except TimeoutException:
                post_dict['views'].append(np.NaN)

            try:
                WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.ID, 'count')))
                num_comments = soup.find_all('h2', id = 'count')[0].get_text()
                num_comments = num_comments.split()[0]
                post_dict['num_comments'].append(num_comments)

            except TimeoutException:
                post_dict['num_comments'].append(np.NaN)
                
            except IndexError:
                post_dict['num_comments'].append(np.NaN)

            try:
                likes = soup.findAll('yt-formatted-string', {"class" : "style-scope ytd-toggle-button-renderer style-text"})[1].get_text()
                post_dict['num_likes'].append(likes)

            except TimeoutException:
                post_dict['num_likes'].append(np.NaN)

            try:
                dislikes = soup.findAll('yt-formatted-string', {"class" : "style-scope ytd-toggle-button-renderer style-text"})[2].get_text()
                post_dict['num_dislikes'].append(dislikes)
            except TimeoutException:
                post_dict['num_dislikes'].append(np.NaN)

            try:
                description = soup.find_all('div', id = 'description')[0].get_text() 
                post_dict['description'].append(description)

            except TimeoutException:
                post_dict['description'].append('')

            post_dict['timestamp_scrape'].append(pd.to_datetime('now'))
            post_dict['date_scrape'].append(dt.date.today())
            post_dict['search_word'].append(search_query)

            driver.back()

        driver.quit()


   
    youtube_data = clean_youtube_data(pd.DataFrame(post_dict))

    youtube_data['channel_subs'] = youtube_data['channel_subs'].replace([''], np.NaN)

    youtube_data['views'] = youtube_data['views'].replace([''], np.NaN)
    youtube_data['num_comments'] = youtube_data['num_comments'].replace([''], np.NaN)

    youtube_data['num_comments_found'] = (youtube_data['num_comments'] > 0)


    return youtube_data

#### The above scrapers have been running since March 2021, they are stored on a AWS server. The below code is collecting the data from the server and cleaning it so it can be used for our project. Since youtube stops in April, we only use the twitter and reddit data from March 2021 to April 2021.

# Initiate Datagrip (PostgreSQL) connection
PASSWORD = "postgres://myoaykflycsnkv:201a55285e96b40152f4865716b479ed9ee7ee4bbfc9ea28cab50cab5b66412c@ec2-54-90-13-87.compute-1.amazonaws.com:5432/d9hthb6b0r501d"
conn = psycopg2.connect(PASSWORD)
conn.autocommit = True
cur = conn.cursor()

#Reddit pull code
sql_reddit_query = """
    SELECT date_scrape, subreddit, AVG(subreddit_subs) AS subreddit_subs, AVG(num_comms) AS num_comms, AVG(score) AS score, AVG(upvote_ratio) AS upvote_ratio, COUNT(id) AS unique_posts
    FROM reddit_df
    GROUP BY subreddit, date_scrape
    ORDER BY date_scrape
    """

reddit_df = pd.read_sql_query(sql_reddit_query, conn)
reddit_df.tail(10)

#saving raw data
reddit_df.to_csv('reddit_data_raw.csv', index=False)

#Twitter pull code
sql_twitter_query = """
    SELECT date_scrape, user_name, AVG(retweet_count) AS retweet_count, AVG(favorite_count) AS favorite_count, AVG(account_followers) AS account_followers, COUNT(tweet_id) AS unique_users
    FROM twitter_df
    GROUP BY user_name, date_scrape
    ORDER BY date_scrape
"""

twitter_df = pd.read_sql_query(sql_twitter_query, conn)
twitter_df.tail(10)

#saving raw data
twitter_df.to_csv('twitter_raw.csv', index=False)

#Youtube Pull Code
sql_youtube_query = """
    SELECT date_scrape, search_word, AVG(channel_subs) AS channel_subs, AVG(views) AS views, AVG(num_comments) AS num_comments, AVG(num_likes) AS num_likes, COUNT(channel_name) AS unique_channels 
    FROM youtube_df
    GROUP BY search_word, date_scrape
    ORDER BY date_scrape
"""

youtube_df = pd.read_sql_query(sql_youtube_query, conn)
youtube_df.tail(10)

#saving raw data
youtube_df.to_csv('youtube_data_raw.csv', index=False)

## Data Cleaning

#Compiling bitcoin data
reddit_bitcoin = filter_df(reddit_df, 'reddit', 'bitcoin', 'BTC-USD')
twitter_bitcoin = filter_df(twitter_df, 'twitter', 'bitcoin', 'BTC-USD')
youtube_bitcoin = filter_df(youtube_df, 'youtube', 'bitcoin', 'BTC-USD')

#Compiling ethereum data
reddit_ethereum = filter_df(reddit_df, 'reddit', 'ethereum', 'ETH-USD')
twitter_ethereum = filter_df(twitter_df, 'twitter', 'ethereum', 'ETH-USD')
youtube_ethereum = filter_df(youtube_df, 'youtube', 'ethereum', 'ETH-USD')

#Compiling cardano data
reddit_cardano = filter_df(reddit_df, 'reddit', 'cardano', 'ADA-USD')
twitter_cardano = filter_df(twitter_df, 'twitter', 'cardano', 'ADA-USD')
youtube_cardano = filter_df(youtube_df, 'youtube', 'cardano', 'ADA-USD')

#Compiling ripple data
reddit_ripple = filter_df(reddit_df, 'reddit', 'ripple', 'XRP-USD')
twitter_ripple = filter_df(twitter_df, 'twitter', 'ripple', 'XRP-USD')
youtube_ripple = filter_df(youtube_df, 'youtube', 'ripple', 'XRP-USD')

#Compiling algorand data
reddit_algorand = filter_df(reddit_df, 'reddit', 'algorand', 'ALGO-USD')
twitter_algorand = filter_df(twitter_df, 'twitter', 'algorand', 'ALGO-USD')
youtube_algorand = filter_df(youtube_df, 'youtube', 'algorand', 'ALGO-USD')

#Compiling dogecoin data
reddit_doge = filter_df(reddit_df, 'reddit', 'dogecoin', 'DOGE-USD')
twitter_doge = filter_df(twitter_df, 'twitter', 'dogecoin', 'DOGE-USD')
youtube_doge = filter_df(youtube_df, 'youtube', 'dogecoin', 'DOGE-USD')

#Compiling chainlink data
reddit_link = filter_df(reddit_df, 'reddit', 'chainlink', 'LINK-USD')
twitter_link = filter_df(twitter_df, 'twitter', 'chainlink', 'LINK-USD')
youtube_link = filter_df(youtube_df, 'youtube', 'chainlink', 'LINK-USD')

#Compiling monero data
reddit_monero = filter_df(reddit_df, 'reddit', 'monero', 'XMR-USD')
twitter_monero = filter_df(twitter_df, 'twitter', 'monero', 'XMR-USD')
youtube_monero = filter_df(youtube_df, 'youtube', 'monero', 'XMR-USD')

#Compiling filecoin data
reddit_filecoin = filter_df(reddit_df, 'reddit', 'filecoin', 'FIL-USD')
twitter_filecoin = filter_df(twitter_df, 'twitter', 'filecoin', 'FIL-USD')
youtube_filecoin = filter_df(youtube_df, 'youtube', 'filecoin', 'FIL-USD')

#Compiling litecoin data
reddit_litecoin = filter_df(reddit_df, 'reddit', 'litecoin', 'LTC-USD')
twitter_litecoin = filter_df(twitter_df, 'twitter', 'litecoin', 'LTC-USD')
youtube_litecoin = filter_df(youtube_df, 'youtube', 'litecoin', 'LTC-USD')

#Compiling tezos data
reddit_tezos = filter_df(reddit_df, 'reddit', 'tezos', 'XTZ-USD')
twitter_tezos = filter_df(twitter_df, 'twitter', 'tezos', 'XTZ-USD')
youtube_tezos = filter_df(youtube_df, 'youtube', 'tezos', 'XTZ-USD')

#Compiling stellar data
reddit_stellar = filter_df(reddit_df, 'reddit', 'stellar', 'XLM-USD')
twitter_stellar = filter_df(twitter_df, 'twitter', 'stellarorg', 'XLM-USD')
youtube_stellar = filter_df(youtube_df, 'youtube', 'stellar crypto', 'XLM-USD')

#creating overall reddit df
reddit_dfs = [reddit_bitcoin, reddit_ethereum, reddit_cardano, reddit_ripple, reddit_algorand, reddit_doge, reddit_link, reddit_monero, reddit_filecoin, reddit_litecoin, reddit_tezos, reddit_stellar]
reddit_clean = pd.concat(reddit_dfs)

#creating overall twitter df
twitter_dfs = [twitter_bitcoin, twitter_ethereum, twitter_cardano, twitter_ripple, twitter_algorand, twitter_doge, twitter_link, twitter_monero, twitter_filecoin, twitter_litecoin, twitter_tezos, twitter_stellar]
twitter_clean = pd.concat(twitter_dfs)

#creating overall youtube df
youtube_dfs = [youtube_bitcoin, youtube_ethereum, youtube_cardano, youtube_ripple, youtube_algorand, youtube_doge, youtube_link, youtube_monero, youtube_filecoin, youtube_litecoin, youtube_tezos, youtube_stellar]
youtube_clean = pd.concat(youtube_dfs)

#creating overall df by combining reddit and youtube
df_final = pd.merge(reddit_clean, youtube_clean, how="right", on="Close")

#adding twitter to final
df_final = pd.merge(df_final, twitter_clean, how="right", on="Close")

#choosing cols we want
cols = ['date_scrape', 'subreddit', 'subreddit_subs', 'num_comms', 'score', 'upvote_ratio', 'unique_posts', 'retweet_count',
       'favorite_count', 'account_followers', 'unique_users', 'channel_subs', 'views', 'num_comments', 'num_likes',
       'unique_channels', 'Close']
df_final = df_final[cols]

#dropping nans
df_final = df_final.dropna()

#renaming columns
df_final = df_final.rename(columns={"subreddit": "coin_name", "subreddit_subs": "r_subscribers", "num_comms": "r_comments",
                        "score" : "r_upvotes", "upvote_ratio" : "r_upvote_ratio", "unique_posts" : "t_unique_posts",
                        "retweet_count" : "t_retweets", "favorite_count" : "t_favorites", "account_followers" : "t_followers",
                        "unique_users" : "t_unique_users", "channel_subs" : "y_subscribers", "views" : "y_views", "num_comments" : "y_comments",
                        "num_likes" : "y_likes", "unique_channels" : "y_unique_channels", "Close" : "price"})

df_final

#saving
df_final.to_csv('crypto_data_clean.csv', index=False)
