==> Code/1 - Exploration.Rmd <==
---
title: "1 - Exploration.R"
---

```{r coin-price-boxplots}
# load libraries
library(kableExtra)                     # for printing tables
library(cowplot)                        # for side by side plots
library(tidyverse)
library(corrplot)                       #For correlation matrix

# read in the train data for exploration
crypto_train = read_csv("/Users/ArpanBagui/Documents/UPenn/Classes/stat-471-final/Data/clean/crypto_train.csv")

# create boxplot of coin prices for each test coin to see distribution
#ethereum
p1 = crypto_train %>% filter(coin_name == "ethereum") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Ethereum Price ($)", 
       y = " ") +
  theme_bw()

#cardano
p2 = crypto_train %>% filter(coin_name == "cardano") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Cardano Price ($)", 
       y = " ") +
  theme_bw()

#ripple
p3 = crypto_train %>% filter(coin_name == "ripple") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Ripple Price ($)", 
       y = " ") +
  theme_bw()

#dogecoin
p4 = crypto_train %>% filter(coin_name == "dogecoin") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Dogecoin Price ($)", 
       y = " ") +
  theme_bw()

#chainlink
p5 = crypto_train %>% filter(coin_name == "chainlink") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Chainlink Price ($)", 
       y = " ") +
  theme_bw()

#monero
p6 = crypto_train %>% filter(coin_name == "monero") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Monero Price ($)", 
       y = " ") +
  theme_bw()

#filecoin
p7 = crypto_train %>% filter(coin_name == "filecoin") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Filecoin Price ($)", 
       y = " ") +
  theme_bw()

#tezos
p8 = crypto_train %>% filter(coin_name == "tezos") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Tezos Price ($)", 
       y = " ") +
  theme_bw()

#stellar
p9 = crypto_train %>% filter(coin_name == "stellar") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Stellar Price ($)", 
       y = " ") +
  theme_bw()

#algorand
p10 = crypto_train %>% filter(coin_name == "algorand") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Algorand Price ($)", 
       y = " ") +
  theme_bw()

#bitcoin
p11 = crypto_train %>% filter(coin_name == "bitcoin") %>%
  ggplot(aes(x = price, y = "price")) +
  geom_boxplot() +
  labs(x = "Bitcoin Price ($)", 
       y = " ") +
  theme_bw()

# combine the plots
price_hists = cowplot::plot_grid(p1, p2, p3, p4, p5, p6, p7, p8, p9, p10, p11, nrow = 3)

# save the boxplots
ggsave(filename = "../Results/coin-price-boxplots.png", 
       plot = price_hists, 
       device = "png", 
       width = 10, 
       height = 6)
```

```{r likes-scatterplots-1}
#examining test coins and their relationship to reddit upvotes, twitter favorites, and youtube likes
#this may give us a feel of which social media has a stronger relationship with price

#starting with first three

#ethereum reddit
p1 = crypto_train %>% filter(coin_name == "ethereum") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/ethereum", 
       y = "Price",
       title = "Ethereum Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#ethereum twitter
p2 = crypto_train %>% filter(coin_name == "ethereum") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Ethereum Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#ethereum youtube
p3 = crypto_train %>% filter(coin_name == "ethereum") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Ethereum Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#cardano reddit
p4 = crypto_train %>% filter(coin_name == "cardano") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/cardano", 
       y = "Price",
       title = "Cardano Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#cardano twitter
p5 = crypto_train %>% filter(coin_name == "cardano") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Cardano Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#cardano youtube
p6 = crypto_train %>% filter(coin_name == "cardano") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Cardano Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#dogecoin reddit
p7 = crypto_train %>% filter(coin_name == "dogecoin") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/dogecoin", 
       y = "Price",
       title = "Dogecoin Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#dogecoin twitter
p8 = crypto_train %>% filter(coin_name == "dogecoin") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Dogecoin Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#dogecoin youtube
p9 = crypto_train %>% filter(coin_name == "dogecoin") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Dogecoin Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

# combine the plots
price_scatterplots = cowplot::plot_grid(p1, p2, p3, p4, p5, p6, p7, p8, p9, nrow = 3)

# save the scatterplots
ggsave(filename = "../Results/coin-price-scatterplots.png", 
       plot = price_scatterplots, 
       device = "png", 
       width = 10, 
       height = 6)
```

```{r likes-scatterplots-2}
#next three coins

#ripple reddit
p1 = crypto_train %>% filter(coin_name == "ripple") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/ripple", 
       y = "Price",
       title = "Ripple Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#ripple twitter
p2 = crypto_train %>% filter(coin_name == "ripple") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Ripple Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#ripple youtube
p3 = crypto_train %>% filter(coin_name == "ripple") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Ripple Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#chainlink reddit
p4 = crypto_train %>% filter(coin_name == "chainlink") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/chainlink", 
       y = "Price",
       title = "Chainlink Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#chainlink twitter
p5 = crypto_train %>% filter(coin_name == "chainlink") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Chainlink Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#chainlink youtube
p6 = crypto_train %>% filter(coin_name == "chainlink") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Chainlink Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#monero reddit
p7 = crypto_train %>% filter(coin_name == "monero") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/monero", 
       y = "Price",
       title = "Monero Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#monero twitter
p8 = crypto_train %>% filter(coin_name == "monero") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Monero Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#monero youtube
p9 = crypto_train %>% filter(coin_name == "monero") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Monero Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

# combine the plots
price_scatterplots = cowplot::plot_grid(p1, p2, p3, p4, p5, p6, p7, p8, p9, nrow = 3)

# save the scatterplots
ggsave(filename = "../Results/coin-price-scatterplots-2.png", 
       plot = price_scatterplots, 
       device = "png", 
       width = 10, 
       height = 6)
```

```{r likes-scatterplots-3}
#final three coins

#filecoin reddit
p1 = crypto_train %>% filter(coin_name == "filecoin") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/filecoin", 
       y = "Price",
       title = "Filecoin Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#filecoin twitter
p2 = crypto_train %>% filter(coin_name == "filecoin") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Filecoin Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#filecoin youtube
p3 = crypto_train %>% filter(coin_name == "filecoin") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Filecoin Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#tezos reddit
p4 = crypto_train %>% filter(coin_name == "tezos") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/tezos", 
       y = "Price",
       title = "Tezos Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#tezos twitter
p5 = crypto_train %>% filter(coin_name == "tezos") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Tezos Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#tezos youtube
p6 = crypto_train %>% filter(coin_name == "tezos") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Tezos Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#stellar reddit
p7 = crypto_train %>% filter(coin_name == "stellar") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/stellar", 
       y = "Price",
       title = "Stellar Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#stellar twitter
p8 = crypto_train %>% filter(coin_name == "stellar") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Stellar Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#stellar youtube
p9 = crypto_train %>% filter(coin_name == "stellar") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Stellar Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#algorand reddit
p10 = crypto_train %>% filter(coin_name == "algorand") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/algorand", 
       y = "Price",
       title = "Algorand Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#algorand twitter
p11 = crypto_train %>% filter(coin_name == "algorand") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Algorand Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#algorand youtube
p12 = crypto_train %>% filter(coin_name == "algorand") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Algorand Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#bitcoin reddit
p13 = crypto_train %>% filter(coin_name == "bitcoin") %>%
  ggplot(aes(x = r_upvotes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "orange") +
  labs(x = "Reddit Upvotes on r/bitcoin", 
       y = "Price",
       title = "Bitcoin Reddit") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#bitcoin twitter
p14 = crypto_train %>% filter(coin_name == "bitcoin") %>%
  ggplot(aes(x = t_favorites, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "light blue") +
  labs(x = "Twitter Favorites on Top Tweets", 
       y = "Price",
       title = "Bitcoin Twitter") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

#bitcoin youtube
p15 = crypto_train %>% filter(coin_name == "bitcoin") %>%
  ggplot(aes(x = y_likes, y = price)) +
  geom_point() + 
  scale_y_log10() +
  geom_smooth(method = "lm", formula = "y~x", se = FALSE, color = "red") +
  labs(x = "Youtube Likes on Top Videos", 
       y = "Price",
       title = "Bitcoin Youtube") +
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5))

# combine the plots
price_scatterplots = cowplot::plot_grid(p1, p2, p3, p4, p5, p6, p7, p8, p9, p10, p11, p12, p13, p14, p15,  nrow = 3)

# save the scatterplots
ggsave(filename = "../Results/coin-price-scatterplots-3.png", 
       plot = price_scatterplots, 
       device = "png", 
       width = 10, 
       height = 6)
```

```{r correlation matrices}
# correlation matrix of Reddit variables
reddit = cor(crypto_train[1:5])
round(reddit, 2)

reddit_matrix = corrplot(reddit, type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45) 

# correlation matrix of Twitter variables
twitter = cor(crypto_train[6:9])
round(twitter, 2)

twitter_matrix = corrplot(twitter, type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45)

# correlation matrix of Youtube variables
youtube = cor(crypto_train[10:14])
round(youtube, 2)

youtube_matrix = corrplot(youtube, type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45)


#create correlation matrix of all variables
res = cor(crypto_train[1:14])
round(res, 2)

all = corrplot(res, type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45)

```
