==> Code/2 - Train-Test-Split.Rmd <==
---
title: "2 - Test-Train-Split"
---

```{r}
# read in the cleaned data
crypto_data = read_csv("/Users/ArpanBagui/Documents/UPenn/Classes/stat-471-final/Data/clean/crypto_data_clean.csv")

set.seed(471) # seed set for reproducibility
n = nrow(crypto_data)
train_samples = sample(1:n, round(0.8*n))
crypto_train = crypto_data %>% filter(row_number() %in% train_samples) 
crypto_test = crypto_data %>% filter(!(row_number() %in% train_samples))

# save the train and test data
write_csv(x = crypto_train, file = "../Data/clean/crypto_train.csv")

write_csv(x = crypto_test, file = "../Data/clean/crypto_test.csv")
```
